added confirmation dialog

+++ src/modules/edc-demo/components/contract-definition-viewer/contract-definition-viewer.component.ts
@@ -2,8 +2,11 @@ import {Component, OnInit} from '@angular/core';
 import {BehaviorSubject, Observable, of} from 'rxjs';
 import {first, map, switchMap} from 'rxjs/operators';
 import {MatDialog} from '@angular/material/dialog';
-import {ContractDefinitionEditorDialog} from '../contract-definition-editor-dialog/contract-definition-editor-dialog.component';
+import {
+  ContractDefinitionEditorDialog
+} from '../contract-definition-editor-dialog/contract-definition-editor-dialog.component';
 import {ContractDefinitionDto, ContractDefinitionService} from "../../../edc-dmgmt-client";
+import {ConfirmationDialogComponent, ConfirmDialogModel} from "../confirmation-dialog/confirmation-dialog.component";
 
 
 @Component({
@@ -26,9 +29,9 @@ export class ContractDefinitionViewerComponent implements OnInit {
         switchMap(() => {
           const contractDefinitions$ = this.contractDefinitionService.getAllContractDefinitions();
           return !!this.searchText ?
-          contractDefinitions$.pipe(map(contractDefinitions => contractDefinitions.filter(contractDefinition => contractDefinition.id.toLowerCase().includes(this.searchText))))
+            contractDefinitions$.pipe(map(contractDefinitions => contractDefinitions.filter(contractDefinition => contractDefinition.id.toLowerCase().includes(this.searchText))))
             :
-          contractDefinitions$;
+            contractDefinitions$;
         }));
   }
 
@@ -37,7 +40,16 @@ export class ContractDefinitionViewerComponent implements OnInit {
   }
 
   onDelete(contractDefinition: ContractDefinitionDto) {
-    this.contractDefinitionService.deleteContractDefinition(contractDefinition.id).subscribe(() => this.fetch$.next(null));
+    const dialogData = ConfirmDialogModel.forDelete("contract definition", contractDefinition.id);
+
+    const ref = this.dialog.open(ConfirmationDialogComponent, {maxWidth: '20%', data: dialogData});
+
+    ref.afterClosed().subscribe(res => {
+      if (res) {
+        this.contractDefinitionService.deleteContractDefinition(contractDefinition.id).subscribe(() => this.fetch$.next(null));
+      }
+    });
+
   }
 
   onCreate() {
